+++ kernel/MAIN.C
@@ -24,25 +24,83 @@
 * @note maximum file size is 32 sectors (16KB)
 * @see c0t.asm
 */
-#include <conio.h>
-
-extern int _heap_start;
+#include <kmem.h>
+#include <conio.h> /* printFormat */
+#include <string.h> /* memset, size_t */
+extern unsigned int _heapStart; /* file: c0t.asm */
 
 void main() {
     #define SIZE 100
     unsigned char *p;
-    //                   size  actual  buffer  null
+    unsigned char far *address = NULL;
+    unsigned int segment = 0;
+    unsigned int offset = 0;
+    char value = 'a';
+    size_t size = 0;
+    int i;
+    /*                   size  actual  buffer  null */
     unsigned char buffer[1+    1+      SIZE+      1 ]={SIZE};
 
-    clearScreen();
-    printFormat("Welcome to NOS, (c)%d By Ahmad Dajani\n", 2020);
+    printFormat("Starting NOS, (c)%d By Ahmad Dajani\n", 2020);
+    initializeMemory(_heapStart);
 
+    /* create MCB */
+    printFormat("Memory test\n");
+    for(i=0; i<9; i++) {
+        address = (unsigned char far *) kmalloc(0xffffL);
+        if(!address) {
+            printFormat("Error: no memory\n");
+            asm hlt;
+        }
+        printFormat("Allocate 0xffff bytes @ %x:%x\n", FP_SEG(address), FP_OFF(address));
+        memset(address, 'a' + i, 0xffff);
+    }
+
+    address = (unsigned char far *) kmalloc(0x98d4);
+    if(!address) {
+        printFormat("\nError: no memory");
+        asm hlt;
+    }
+    printFormat("Allocate 0x98d4 @ %x:%x\n", FP_SEG(address), FP_OFF(address));
+    memset(address, '$', 0x98d4);
+
+    /* Free */
+    printFormat("Free address %x:%x\n", FP_SEG(address), FP_OFF(address));
+    kfree(address);
+
+    address = (unsigned char far *) kmalloc(0x200);
+    if(!address) {
+        printFormat("\nError: no memory");
+        asm hlt;
+    }
+    printFormat("Allocate 0x200 @ %x:%x\n", FP_SEG(address), FP_OFF(address));
+    memset(address, '!', 0x200);
+
+    /* Main memory tester */
     while(1) {
-        printFormat("%%");
+        printFormat("\nEnter segment in hex:");
         p = readString(buffer);
-        printFormat("\n\t your command is [%s]\n", p);
+        segment = convertHexStringToInteger(p);
+
+        printFormat("\nEnter offset in hex:");
+        p = readString(buffer);
+        offset = convertHexStringToInteger(p);
+
+        printFormat("\nEnter value in hex:");
+        p = readString(buffer);
+        value = convertHexStringToInteger(p);
+
+        printFormat("\nEnter size in hex:");
+        p = readString(buffer);
+        size = convertHexStringToInteger(p);
+
+        address = MK_FP(segment, offset);
+        memset(address, value, size);
+        printFormat("\n-------> DONE");
+
         if(p[0]=='x') break;
     }
     printFormat("\nBye:)\n");
+
     asm hlt
 }